Add a cache mechanism to gulp
Running `gulp --tasks-simple` is a bit slow (around 0.75s on my
computer). This patch caches the command's result so the command doesn't
have to be executed each time the user wants the task list. To
invalidate the cache, the user must pass a prefix argument to
`gulp` (i.e., `C-u M-x gulp`).