Lookup gulp binary in current buffer
The gulp binary could be installed as a project local NPM dependency and
will therefor not be present on the global `exec-path`.

Instead we search for the executable when the user runs the `gulp`
command and we are still in the current buffer. This will then be able
to locate the gulp binary if the node_modules bin folder has been added
to the `exec-path` in that buffer (i.e. by the help of
`add-node-modules-path`)